+++ lib/plug/accesslog/default_formatter/request_serving_time.ex
@@ -8,7 +8,11 @@ defmodule Plug.AccessLog.DefaultFormatter.RequestServingTime do
   @doc """
   Appends to log output.
   """
-  @spec append(String.t, Plug.Conn.t, :secs | :msecs | :usecs) :: String.t
+  @spec append(String.t, Plug.Conn.t, String.t | :secs | :msecs | :usecs) :: String.t
+  def append(message, conn, "s"),  do: append(message, conn, :secs)
+  def append(message, conn, "ms"), do: append(message, conn, :msecs)
+  def append(message, conn, "us"), do: append(message, conn, :usecs)
+
   def append(message, conn, format), do: message <> serving_time(conn, format)
 
   defp serving_time(conn, format) do