+++ test/plug/accesslog/default_formatter/request_serving_time_test.exs
@@ -6,7 +6,7 @@ defmodule Plug.AccessLog.DefaultFormatter.RequestServingTimeTest do
 
   alias Plug.AccessLog.DefaultFormatter
 
-  test "%D" do
+  test "%D == %{us}T" do
     timestamp = :os.timestamp()
     conn      =
          conn(:get, "/")
@@ -15,11 +15,13 @@ defmodule Plug.AccessLog.DefaultFormatter.RequestServingTimeTest do
     # relying on microsecond testing precision is flaky.
     # check done using "microseconds taken < 1 second"
     { usecs, _ } = DefaultFormatter.format("%D", conn) |> Integer.parse()
+    { t_var, _ } = DefaultFormatter.format("%{us}T", conn) |> Integer.parse()
 
     assert usecs < 1000000
+    assert t_var < 1000000
   end
 
-  test "%M" do
+  test "%M == %{ms}T" do
     timestamp = :os.timestamp()
     conn      =
          conn(:get, "/")
@@ -28,17 +30,20 @@ defmodule Plug.AccessLog.DefaultFormatter.RequestServingTimeTest do
     # relying on millisecond testing precision is flaky.
     # check done using "milliseconds taken < 1 second"
     { msecs, _ } = DefaultFormatter.format("%M", conn) |> Integer.parse()
+    { t_var, _ } = DefaultFormatter.format("%{ms}T", conn) |> Integer.parse()
 
     assert msecs < 1000
+    assert t_var < 1000
   end
 
-  test "%T" do
+  test "%T == %{s}T" do
     timestamp = :os.timestamp()
     conn      =
          conn(:get, "/")
       |> put_private(:plug_accesslog, %{ timestamp: timestamp })
 
     assert "0" == DefaultFormatter.format("%T", conn)
+    assert "0" == DefaultFormatter.format("%{s}T", conn)
   end
 
   test "%T rounding" do
@@ -48,5 +53,6 @@ defmodule Plug.AccessLog.DefaultFormatter.RequestServingTimeTest do
       |> put_private(:plug_accesslog, %{ timestamp: timestamp })
 
     assert "1" == DefaultFormatter.format("%T", conn)
+    assert "1" == DefaultFormatter.format("%{s}T", conn)
   end
 end